requirements.txt
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
OpenAI
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
langchain-classic
faiss-cpu
unstructured
libmagic